+++ src/vi/la/cpu/cpu_matrix.h
@@ -11,7 +11,7 @@ namespace cpu {
 /// Matrix implementation for CPU context
 class matrix : public vi::la::matrix_implementation {
 public:
-  matrix(cpu_context& context, size_t rows, size_t columns, const double* initial_values);
+  matrix(cpu_context& context, size_t rows, size_t columns, const float* initial_values);
   virtual ~matrix();
 
   size_t row_count() const;
@@ -19,15 +19,15 @@ public:
 
   virtual vi::la::context& owning_context() const;
 
-  virtual double* raw_data();
+  virtual float * raw_data();
 
-  double* get();
+  float* get();
 
 private:
   cpu_context& _context;
   size_t _row_count;
   size_t _column_count;
-  double* _buffer;
+  float* _buffer;
 };
 }
 }
